+++ frontend/src/features/Mission/components/MissionForm/ActionList/styles.ts
@@ -1,0 +1,27 @@
+import styled from 'styled-components'
+
+export const ActionLabel = styled.div`
+  display: flex;
+  flex-grow: 1;
+
+  > .Element-IconBox {
+    margin-right: 8px;
+  }
+
+  > p {
+    margin-top: 0px;
+    color: ${p => p.theme.color.gunMetal};
+    padding: 1px 0px 0 0;
+    min-height: 26px;
+  }
+`
+
+export const Head = styled.div`
+  align-items: flex-start;
+  display: flex;
+
+  /* TODO Remove the padding if iconSize is set in monitor-ui. */
+  > button {
+    padding: 0;
+  }
+`
